face-api/requirements.txt
aniso8601==9.0.1
autopep8==1.6.0
click==8.1.3
dlib==19.23.1
Flask==2.2.2
Flask-RESTful==0.3.9
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
Pillow==9.3.0
pycodestyle==2.8.0
python-dotenv==0.21.0
pytz==2022.7
six==1.16.0
toml==0.10.2
Werkzeug==2.2.2
zipp==3.11.0
